+++ src/composables/useFieldMapping.js
@@ -63,6 +63,7 @@ export function useFieldMapping() {
     const keys = Object.keys(selectedRawItem).filter((key) => {
       if (key === 'scan') return false
       if (key === 'suspendEditing') return false
+      if (key === '__onlineMeta') return false
       if (!appliedShowOnlyNonEmptyFields.value) return true
       if (isAutosuggestFieldType(appliedUserConfigFields.value[key]?.type)) return true
       return !isFieldValueEmpty(selectedRawItem[key])